requirements.txt
folium==0.20.0
haversine==2.9.0
numpy==2.3.4
pandas==2.3.3
Pillow==12.0.0
plotly==6.3.1
streamlit==1.50.0
streamlit_folium==0.25.3
